refactor: Extract BookmarksSectionItems component from BookmarksSection
